Candies
Abhi is a salesman. He was given two types of candies, which he is selling in NN different cities.
For the prices of the candies to be valid, Abhi's boss laid down the following condition:
A given type of candy must have distinct prices in all NN cities.
In his excitement, Abhi wrote down the prices of both the candies on the same page and in random order instead of writing them on different pages. Now he is asking for your help to find out if the prices he wrote are valid or not.
You are given an array AA of size 2N2N. Find out whether it is possible to split AA into two arrays, each of length NN, such that both arrays consist of distinct elements.
Both arrays can have distinct elements only if no element in the original array is repeated more than twice