Chef Diet
Chef decided to go on a diet during the following NN days (numbered 11 through NN). Part of the diet plan is to eat KK grams of protein during each day. For each valid ii, Chef wants to buy AiAi​ grams of protein in the morning of the ii-th day and then eat KK grams of protein as part of his dinner. If he has any protein remaining, he can store it and use it in later dinners. Initially, Chef is storing 00 grams of protein.
Determine whether Chef will have enough protein all the time during his diet. In case he will not have enough, find the first day on which Chef will be unable to eat KK grams of protein.