Break the Stick
Chef has a stick of length NN.

He can break the stick into 22 or more parts such that the parity of length of each part is same. For example, a stick of length 1111 can be broken into three sticks of lengths {3,3,5}{3,3,5} since each part is odd, but it cannot be broken into two sticks of lengths {5,6}{5,6} since one is even and the other is odd.

Chef can then continue applying this operation on the smaller sticks he obtains, as many times as he likes.

Can Chef obtain a stick of length exactly XX by doing this?